+++ src/main/java/org/example/springordermanagement/error/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package org.example.springordermanagement;
+package org.example.springordermanagement.error;
 
 public class ResourceNotFoundException extends Exception {
     public ResourceNotFoundException(String message) {
