+++ src/main/java/org/example/springordermanagement/order/OrderRepository.java
@@ -1,4 +1,4 @@
-package org.example.springordermanagement;
+package org.example.springordermanagement.order;
 
 import org.springframework.data.jpa.repository.JpaRepository;
